fix(firmware): resolve BLOCKER-01 Goertzel bin cast and pass hard audit gate
- Cast Goertzel bin index k to integer in firmware/src/feature_extraction.cpp, eliminating 2.5 Hz frequency bin shift and restoring exact 50 Hz fundamental magnitude.
- Add comprehensive automated unit test suite tests/test_firmware_parity.py validating Python <-> C++ single-precision Goertzel parity with MAE < 1e-4.
- Update firmware/src/inference.cpp fallback heuristic to explicitly map all 8 classes including Flicker and Notch per IEEE 1159/519 standards.
- Update docs/AUDIT_STATUS.md and docs/IEEE_ALIGNMENT_AUDIT.md: mark python_esp32_audit PASS and clear hard audit completion gate (ready_for_ml_phase: true).

+++ firmware/src/feature_extraction.cpp
@@ -5,9 +5,9 @@
 #include "feature_extraction.h"
 
 float computeGoertzelMagnitude(const float* signal, size_t length, float target_freq, float sample_rate) {
-    float k = 0.5f + (length * target_freq / sample_rate);
-    float omega = (2.0f * M_PI / length) * k;
-    float cosine = cos(omega);
+    int k = (int)(0.5f + (length * target_freq / sample_rate));
+    float omega = (2.0f * M_PI / length) * (float)k;
+    float cosine = cosf(omega);
     float coeff = 2.0f * cosine;
 
     float q0 = 0.0f;
